knn-colours/pipeline/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via httpcore
boto3==1.24.92
    # via -r requirements.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
elastic-transport==8.4.0
    # via elasticsearch
elasticsearch==8.5.1
    # via -r requirements.in
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
imageio==2.28.1
    # via scikit-image
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.2.0
    # via scikit-learn
lazy-loader==0.2
    # via scikit-image
loguru==0.6.0
    # via -r requirements.in
networkx==3.1
    # via scikit-image
numpy==1.24.3
    # via
    #   imageio
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
packaging==23.1
    # via scikit-image
pillow==9.5.0
    # via
    #   -r requirements.in
    #   imageio
    #   scikit-image
python-dateutil==2.8.2
    # via botocore
pywavelets==1.4.1
    # via scikit-image
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.6.0
    # via boto3
scikit-image==0.20.0
    # via -r requirements.in
scikit-learn==1.2.2
    # via -r requirements.in
scipy==1.10.1
    # via
    #   scikit-image
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2023.4.12
    # via scikit-image
tqdm==4.65.0
    # via -r requirements.in
urllib3==1.26.14
    # via
    #   botocore
    #   elastic-transport
